refactor(sync): drop post-merge-validator re-export shim
Migrate 11 sync test files and the sync-engine consumer to import
per-validator helpers (normalizeFrontingSessions, detectHierarchyCycles,
ENTITY_FIELD_MAP, getEntityTypeByFieldName, etc.) directly from
packages/sync/src/validators/<name>.js instead of through the
post-merge-validator orchestrator. Remove the 16-line re-export block
that previously aliased the validator surface for the smaller-diff
split-PR. runAllValidations remains the public orchestrator entry.

Two tests use `import *` namespace imports for vi.spyOn on
runAllValidations — those legitimately access the orchestrator and
keep the namespace form unchanged.

Pre-production policy: no aliases, shims, or re-exports.
