+++ packages/sync/src/__tests__/post-merge-validator-fronting-sessions.test.ts
@@ -14,8 +14,9 @@ import { WasmSodiumAdapter } from "@pluralscape/crypto/wasm";
 import { beforeAll, beforeEach, describe, expect, it } from "vitest";
 
 import { createFrontingDocument, createSystemCoreDocument } from "../factories/document-factory.js";
-import { normalizeFrontingSessions, runAllValidations } from "../post-merge-validator.js";
+import { runAllValidations } from "../post-merge-validator.js";
 import { EncryptedSyncSession } from "../sync-session.js";
+import { normalizeFrontingSessions } from "../validators/fronting.js";
 
 import {
   makeFrontingSession,
