+++ packages/sync/src/__tests__/post-merge-validator-m4-webhook-config.test.ts
@@ -11,7 +11,8 @@
 import { WasmSodiumAdapter } from "@pluralscape/crypto/wasm";
 import { beforeAll, beforeEach, describe, expect, it } from "vitest";
 
-import { normalizeWebhookConfigs, runAllValidations } from "../post-merge-validator.js";
+import { runAllValidations } from "../post-merge-validator.js";
+import { normalizeWebhookConfigs } from "../validators/webhook-config.js";
 
 import {
   makeKeys,
